completed CRUD todos with JWT authentication

+++ src/udemy/fastapi/project3/todoapp/routers/todos.py
@@ -36,13 +36,18 @@ class TodoRequest(BaseModel):
 
 #Depends : dependency injection
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return  db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    return  db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model =  db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+    todo_model =  (db.query(Todos).filter(Todos.id == todo_id)
+                   .filter(Todos.id == user.get('id')).first())
     if todo_model is not None:
         return todo_model
 
@@ -58,11 +63,16 @@ async def create_todo(user: user_dependency, db: db_dependency, todo_request: To
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency,
+async def update_todo(user: user_dependency,
+                      db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int= Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
 
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id'))
+                  .first())
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found')
     todo_model.title = todo_request.title
@@ -74,8 +84,15 @@ async def update_todo(db: db_dependency,
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency,
+                      db: db_dependency,
+                      todo_id: int=Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = (db.query(Todos).filter(Todos.id == todo_id)
+                  .filter(Todos.owner_id == user.get('id'))
+                  .first())
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Item not found')
     db.query(Todos).filter(Todos.id == todo_id).delete()
